2026.08.01

+++ 00.image_tokenizer_training/fsq_training/FSQ_8_codebook_16K_resize256_inference/inference_config.py
@@ -27,7 +27,7 @@ class config:
 
     use_amp = False
     # torch.float16 or torch.bfloat16
-    amp_type = torch.float16
+    amp_type = torch.bfloat16
 
     inference_image_dir = './test_images'
     save_image_dir = './test_images_result'
